+++ ProjektCpp2/CDroga.h
@@ -7,7 +7,7 @@ class CDroga
 private:
 	int ID;
 	int id_Sk[2];
-	int dlugosc;
+	double dlugosc;
 	double kat;
 	double sinus;
 	double cosinus;
@@ -15,11 +15,11 @@ private:
 	vector <int> pojazdyK1;
 	vector <int> pojazdyK2;
 public:
-	CDroga(int id, int sk1, int sk2, int dl, double s, double c);
+	CDroga(int id, int sk1, int sk2, double dl, double s, double c);
 	int getID();
-	int getDlugosc();
+	double getDlugosc();
 	int getIdSk(int i);
-	float getKat();
+	double getKat();
 	double getSinus();
 	double getCosinus();
 	int getSzerokosc();
